% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgit.R
\name{download}
\alias{download}
\title{Download GitHub/Gitee repo release or archive file}
\usage{
download(repo, destdir, release = NULL, gitee = FALSE, ...)
}
\arguments{
\item{repo}{A GitHub/Gitee repo in the format of \code{username/repo}, e.g. \code{ShixiangWang/tinyscholar}.}

\item{destdir}{A target path to save the file.}

\item{release}{Set to the version number (e.g. \code{v1.0}) for downloading release file.}

\item{gitee}{If \code{TRUE}, download from Gitee repo instead of GitHub repo.}

\item{...}{Other arguments passing to \link[utils:download.file]{utils::download.file}.}
}
\value{
Nothing
}
\description{
Download GitHub/Gitee repo release or archive file
}
\examples{
\donttest{
x <- tempdir()
download("ShixiangWang/tinyscholar", destdir = x)
dir(x)
}
}
