% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgit.R
\name{install}
\alias{install}
\title{Install R packages}
\usage{
install(pkg, gitee = FALSE, ...)
}
\arguments{
\item{pkg}{A package name for Git(hub) package (e.g. \code{ShixiangWang/ezcox}) or
a list of packages (e.g. \code{c("dplyr", "maftools")}) for normal packages.}

\item{gitee}{If \code{TRUE}, install package from gitee.}

\item{...}{Other arguments passing to \link[remotes:install_git]{remotes::install_git},
\link[BiocManager:install]{BiocManager::install} or \link[remotes:install_github]{remotes::install_github} based on input.}
}
\value{
Nothing.
}
\description{
Install R packages
}
\examples{
\dontrun{
install("ShixiangWang/ezcox")
install("ShixiangWang/tinyscholar", gitee = TRUE)
install(c("ggplot2", "Biobase"))
}
}
