% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{rm_paths}
\alias{rm_paths}
\title{Remove paths (files/directories) from your system}
\usage{
rm_paths(paths)
}
\arguments{
\item{paths}{A character vector with the names of the file(s) or directories to be deleted.}
}
\value{
A logical value indicating remove status.
}
\description{
Remove paths (files/directories) from your system
}
\examples{
x <- tempfile()
file.create(x)
yes <- rm_paths(x)
print(yes)
}
