% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfun-package.R
\docType{package}
\name{wfun-package}
\alias{wfun}
\alias{wfun-package}
\title{wfun: Miscellaneous Functions by 'Shixiang Wang'}
\description{
Miscellaneous useful functions commonly used by
    'Shixiang Wang'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShixiangWang/wfun}
  \item Report bugs at \url{https://github.com/ShixiangWang/wfun/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
