% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgit.R
\name{clone}
\alias{clone}
\title{Clone a Git repository}
\usage{
clone(url, local_path, gitee = FALSE, reset_remote = FALSE, ...)
}
\arguments{
\item{url}{The remote repository to clone}

\item{local_path}{Local directory to clone to.}

\item{gitee}{If \code{TRUE}, clone repository from gitee.}

\item{reset_remote}{if \code{TRUE}, reset GitHub repository remote url.}

\item{...}{Other arguments passing to \link[git2r:clone]{git2r::clone}}
}
\value{
Nothing
}
\description{
Clone a Git repository
}
\examples{
\dontrun{
 clone("ShixiangWang/ezcox", file.path(tempdir(), "ezcox"), reset_remote = TRUE)
 clone("ShixiangWang/tinyscholar", file.path(tempdir(), "ezcox"), gitee = TRUE)
}
}
